refactor(gatekeep): modularizar security_check em submódulos focados
