+++ gatekeep/__init__.py
@@ -1,0 +1,1 @@
+# Pacote gatekeep contendo submódulos de segurança e revisão de código
